Move string comparison tests into their own class.
