+++ test/EFCore.MySql.FunctionalTests/Query/CaseSensitiveNorthwindQueryMySqlFixture.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Infrastructure;
+﻿using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.TestUtilities;
 using Pomelo.EntityFrameworkCore.MySql.FunctionalTests.TestUtilities;
 
@@ -10,12 +9,5 @@ namespace Pomelo.EntityFrameworkCore.MySql.FunctionalTests.Query
     {
         protected override string StoreName => "NorthwindCs";
         protected override ITestStoreFactory TestStoreFactory => MySqlNorthwindTestStoreFactory.InstanceCs;
-
-        public override DbContextOptionsBuilder AddOptions(DbContextOptionsBuilder builder)
-        {
-            var optionsBuilder = base.AddOptions(builder);
-            new MySqlDbContextOptionsBuilder(optionsBuilder).EnableStringComparisonTranslations();
-            return optionsBuilder;
-        }
     }
 }
